% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/background.R
\name{background}
\alias{background}
\title{Background Info}
\usage{
background(image.jpg = NA)
}
\arguments{
\item{image.jpg}{Image name in the directory, in .jpg extension created in an image editor.}
}
\value{
Height and Width dimensions from background.
}
\description{
\code{background} is used to import an image.jpg previously created in
an image editor that corresponds to a video frame used in Swistrack4
that does not contain the tracked particle. Later, this file will
be used to define the zones of interest to be analyzed.
}
\examples{
background(image.jpg = system.file('extdata', 'bg-swistrack-1.jpg', package = 'swistrackr'))

}
