% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outputswtr-data.R
\docType{data}
\name{output.swtr}
\alias{output.swtr}
\title{Exemplo de Coordenadas de rastreamento}
\format{
Uma tabela com 3 colunas, frame, x e y
}
\source{
Valores gerados no Swistrack4
}
\usage{
output.swtr
}
\description{
Um dataset  contendo dados de coordenadas de um objeto.
}
\keyword{datasets}
