% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/times.R
\name{times}
\alias{times}
\title{Calculate time in zones}
\usage{
times(output.swtr = NA, zones.swtr = NA, video.swtr = NA)
}
\arguments{
\item{output.swtr}{Swistrack4 output containing the xy coordinates
and frames of the video tracked particle}

\item{zones.swtr}{Table with the coordinates of the zones of interest
created with the \code{zones} function}

\item{video.swtr}{Object containing the information of the video
created with the \code{track_video} function}
}
\value{
A table containing the values of time that the particle
remained in each of the zones of interest.
}
\description{
\code{times} calculates the time in seconds that the tracked particle
remains in the zones of interest. The function uses the framerate
of the analyzed video and the number of frames located within the
zone of interest as parameters to calculate the result.
}
\examples{
times(output.swtr, zones.swtr, video.swtr)

}
