% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/track_video.R
\name{track_video}
\alias{track_video}
\title{Video Analyzed Info}
\usage{
track_video(video.avi = NA)
}
\arguments{
\item{video.avi}{Video name in the directory, in .avi extension analyzed in Swistrack4 program.}
}
\value{
Running time in seconds and framerate from video
}
\description{
\code{track_video} is used to read a video in .avi extension that has been
analysed in the Swistrack4. Subsequently, this file will be used to
measure the running time in seconds that a particle of interest remained
in the user's zones of interest.
}
\examples{
track_video(video.avi = system.file('extdata', 'video-swistrack-1.avi', package = 'swistrackr'))

}
