% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/videoswtr-data.R
\docType{data}
\name{video.swtr}
\alias{video.swtr}
\title{Exemplo de objeto com informacoes de um video analisado}
\format{
Uma tabela com 3 colunas, Nome do video, tempo e framerate
}
\source{
Valores gerados a partir de um video
}
\usage{
video.swtr
}
\description{
Um objeto contendo informacoes de video contendo tempo e framerate.
}
\keyword{datasets}
