% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zones.R
\name{zones}
\alias{zones}
\title{Define Zones of Interest}
\usage{
zones(image.jpg = NA, nzones = 1, pzones = 4)
}
\arguments{
\item{image.jpg}{Background image in .jpg format}

\item{nzones}{Number of interest zones in the video}

\item{pzones}{Number of points to select the zone of interest}
}
\value{
A table containing the x and y axis values for each zone of interest.
}
\description{
With \code{zones} the user determines the zones of interest, using
the background.jpg previously created. Zones of interest are used
to calculate information such as time in seconds, number of visits, or
order of zones visited.
}
\examples{
zones('image.jpg = bg.image', nzones = 2, pzones = 4)

}
