% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zonesswtr-data.R
\docType{data}
\name{zones.swtr}
\alias{zones.swtr}
\title{Exemplo de Coordenadas para zonas de interesse}
\format{
Uma tabela com 3 colunas, Zones, x e y
}
\source{
Valores que seriam gerados pelo usuario manualmente.
}
\usage{
zones.swtr
}
\description{
Um dataset contendo coordenadas para marcar zonas de interesse.
}
\keyword{datasets}
